config/webserver/emmental_flask/requirements.txt
Flask
mysql_connector
bcrypt
flask-login
